arm64: dts: qcom: msm8916: Add blsp_i2c1
MSM8916 has another I2C QUP controller that can be enabled on
GPIO 2 and 3.

Add blsp_i2c1 to msm8916.dtsi and disable it by default.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
On samsung-a2015 there is an external TFA9895 speaker amplifier
connected to blsp_i2c1. There isn't a proper mainline driver
for it (yet?), but it's nice to have the I2C bus in mainline at least.