dt-bindings: vendor-prefixes: Add yet another for ST-Ericsson
Unfortunately the vendor prefix for ST-Ericsson is used very
inconsistently. "ste," and "stericsson," are already documented,
but some things in the kernel use "st-ericsson," which is not
documented yet.

st-ericsson,u8500 is documented in bindings/arm/ux500/boards.txt,
and is used to match the machine code and the generic DT cpufreq
driver.

Add it to the list of vendor prefixes.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Changes in v2: none, added new patch to deprecate other vendor prefixes
v1: https://lore.kernel.org/linux-devicetree/20191120121720.72845-1-stephan@gerhold.net/