media: venus: hfi_parser: Ignore HEVC encoding for V1
Some older MSM8916 Venus firmware versions also seem to indicate
support for encoding HEVC, even though they really can't.
This will lead to errors later because hfi_session_init() fails
in this case.

HEVC is already ignored for "dec_codecs", so add the same for
"enc_codecs" to make these old firmware versions work correctly.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
This makes Venus work on the Samsung Galaxy S4 Mini VE (MSM8916),
which is stuck on a rather old Venus firmware version (1.6-00040).
The firmware is signed, so unfortunately I'm not aware of a way to upgrade it...