firmware: qcom: scm: Add support for MC boot address API
It looks like the old QCOM_SCM_BOOT_SET_ADDR API is broken on some
MSM8916 firmware versions that implement the newer SMC32 calling
convention. It just returns -EINVAL no matter which arguments are
being passed.

This does not cause any problems downstream because it first tries
to use the new multi-cluster API replacement which is working fine.

Implement support for the multi-cluster variant of the SCM call
by attempting it first but still fallback to the old call in case
of an error. Also, to be absolutely sure only use the multi-cluster
variant with the SMC calling convention since older platforms should
not need this.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Changes in v2: None.

The diff generated by Git is a bit hard to read sadly, what I did
essentially is:

  1. Add __qcom_scm_set_boot_addr_mc()
  2. Rename original qcom_scm_set_cold/warm_boot_addr() to
     static __qcom_scm_set_cold/warm_boot_addr()
  3. Make new qcom_scm_set_cold/warm_boot_addr() call
     __qcom_scm_set_boot_addr_mc() first and then fall back to the old
     __qcom_scm_set_cold/warm_boot_addr() as before