arm64: dts: qcom: msm8916: Simplify pinctrl configuration
So far we have been separating pinctrl entries into pinmux/pinconf.
It turns out it is also possible to combine them: The advantage is
that the device tree is overall more concise because the "pins"
to configure just need to be specified once, not separately for
pinmux/pinconf.

Using the simpler form only for new entries would be rather confusing.
This commit makes all MSM8916 device trees use the simplfied form.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Bjorn suggested using this form here:
https://lore.kernel.org/linux-arm-msm/20200514172326.GC279327@builder.lan/