arm64: dts: qcom: msm8916: Replace invalid bias-pull-none property
msm8916-pins.dtsi specifies "bias-pull-none" for most of the audio
pin configurations. This was likely copied from the qcom kernel fork
where the same property was used for these audio pins.

However, "bias-pull-none" actually does not exist at all - not in
mainline and not in downstream. I can only guess that the original
intention was to configure "no pull", i.e. bias-disable.

Change it to that instead.

Fixes: 143bb9ad85b7 ("arm64: dts: qcom: add audio pinctrls")
Cc: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
The reason why the qcom kernel fork used "bias-pull-none" instead
of "bias-disable" there is still very unclear to me. That property
is used only for some audio pins, and seemingly only on MSM8916/MSM8939.

Very confusing :)