arm64: dts: qcom: msm8916: Pull down PDM GPIOs during sleep
The original qcom kernel changed the PDM GPIOs to be pull-down
during sleep at some point. Reportedly this was done because
there was some "leakage at PDM outputs during sleep":

  https://source.codeaurora.org/quic/la/kernel/msm-3.10/commit/?id=0f87e08c1cd3e6484a6f7fb3e74e37340bdcdee0

I cannot say how effective this is, but everything seems to work
fine with this change so let's apply the same to mainline just
to be sure.

Cc: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
I found this at some point while trying to make audio work
(searched for differences in mainline compared to downstream).
This was not a difference that prevented it to work,
but I thought I would just submit it in case this could improve
power efficiency a bit.

Feel free to ignore this if this is stupid.