hwspinlock: qcom: Drop potentially incorrect max_register
At least on some of the "older" modern Qualcomm platforms the memory
region for hwspinlock only has a size of 0x20000 (for 32 locks with
a stride of 0x1000). In that case, setting a max_register = 0x40000
(as used by newer SoCs) is wrong.

It does not cause any functional problems, but given that max_register
is optional, we might as well drop it entirely.

Fixes: 7a1e6fb1c606 ("hwspinlock: qcom: Allow mmio usage in addition to syscon")
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>