ASoC: msm8916-wcd-analog: Add outputs to connect external speaker PA
Some MSM8916 smartphones have an analog speaker amplifier connected to
the HPHR output. This can be modelled as auxiliary device specified in
the device tree.

However, one problem remains: Since HPHR is used for both the headphones
and the speaker, connecting the speaker amplifier directly to the HPHR
output on the DAPM graph causes it to get enabled every time the
headphones are enabled.

To avoid this we need an extra switch, so that we need to enable a
mixer before the speaker gets enabled. This commit adds an extra
"HPHL/R Ext PA" output and "HPHL/R Ext" mixer that allows enabling
the headphones and speaker separately in this situation.

Cc: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>