clk: qcom: smd: Mark "rate" clocks as CLK_IGNORE_UNUSED
Some of the clocks defined with DEFINE_CLK_SMD_RPM() are used to scale
interconnects. If the interconnect driver is missing/disabled we don't want
to disable these even when unused, because otherwise the system will hang.

So far this did not cause any problems because we did not make the clk
core aware of the fact that we force on all clocks during "handoff".
Therefore, unused clocks just stayed always-on.

To prepare for disabling unused branch clocks at least, mark all the "rate"
clocks as CLK_IGNORE_UNUSED, so we don't end up disabling the interconnect
clocks if the interconnect driver is missing/disabled.

Eventually this could be optimized further by removing the flag from
some of the other clocks that are not used for interconnect.
However, that needs more validation on other platforms first.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>