iio: imu: inv_mpu6050: Add support for MPU-6880
MPU-6880 seems to be very similar to MPU-6500 / MPU-6050 and it works
fine with some minor additions for the mpu6050 driver.

Add the necessary defines for it and make it use the same registers as
MPU-6500 but with a FIFO size of 4096.

Cc: Jean-Baptiste Maneyrol <jmaneyrol@invensense.com>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
I tested this on a Xiaomi Redmi 2 (or Wingtech WT88047).

Unfortunately I was not able to find a proper register description
for MPU-6880. :( It seems to work fine with both 6050 or 6500 registers
but I guess that it is more similar to MPU-6500. Not sure though.

I'm mainly unsure about
	.accl_offset		= INV_MPU6500_REG_ACCEL_OFFSET,
vs
	.accl_offset		= INV_MPU6050_REG_ACCEL_OFFSET,

Both registers read some funny values but I cannot say which one
is more appropriate...

Perhaps Jean-Baptiste can shed some light on this? :)