dt-bindings: iio: imu: mpu6050: Document invensense,mpu6880
MPU-6880 seems to be very similar to MPU-6500 / MPU-6050 and it works
fine with some minor additions for the mpu6050 driver.
Add a compatible for it to the binding documentation.

Cc: Jean-Baptiste Maneyrol <jmaneyrol@invensense.com>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
---
Note: This applies only on top of the patch for the YAML conversion:
https://lore.kernel.org/linux-iio/20201128173343.390165-3-jic23@kernel.org/
I guess that one will be applied very soon so I think this is easier :)